Done with the project, now just need to start testing clear

+++ controllers/clothingItem.js
@@ -66,9 +66,17 @@ const likeItem = (req, res) =>
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
 
-  ClothingItem.findByIdAndDelete(itemId)
+  ClothingItem.findById(itemId)
     .orFail()
-    .then((item) => res.status(200).send({ data: item }))
+    .then((item) => {
+      if (item.owner.toString() !== req.user._id) {
+        return res.status(403).send({ message: "You cannot delete this item" });
+      }
+
+      return ClothingItem.findByIdAndDelete(itemId)
+        .orFail()
+        .then((item) => res.status(200).send({ data: item }));
+    })
     .catch((err) => {
       console.error(err.name);
       console.error(err.message);
